Added POC of folders, added abstract class song
